requirements.txt
# Flask Backend
flask==2.0.3
werkzeug==2.0.3
flask-cors==3.0.10
python-dotenv==1.0.0

# Streamlit Frontend
streamlit==1.22.0
requests==2.28.2

# Google Gemini API
google-generativeai==0.3.0

# Utilities
python-dotenv==1.0.0